+++ lib/SILOptimizer/LoopTransforms/CMakeLists.txt
@@ -1,5 +1,6 @@
 silopt_register_sources(
   ArrayBoundsCheckOpts.cpp
+  ArrayPropertyOpt.cpp
   COWArrayOpt.cpp
   LoopRotate.cpp
   LoopUnroll.cpp
