+++ test/SILOptimizer/array_specialize.sil
@@ -1,4 +1,4 @@
-// RUN: %target-sil-opt -enable-sil-verify-all %s -array-specialize
+// RUN: %target-sil-opt -enable-sil-verify-all %s -array-property-opt
 
 sil_stage canonical
